Restaurant_Order_and_Customer/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.0